refactor: move utils to Calendar/DatePicker
