+++ packages/components/src/Calendar/enums.ts
@@ -11,3 +11,12 @@ export const DayOfWeek = {
 // Intentionally naming the type the same as the const for tree-shaking
 // eslint-disable-next-line @typescript-eslint/no-redeclare
 export type DayOfWeek = (typeof DayOfWeek)[keyof typeof DayOfWeek]
+
+export const DateFormat = {
+  Numeral: "P", // eg. 01/15/2022
+  Text: "PP", // eg. Jan 15, 2022
+} as const
+
+// Intentionally naming the type the same as the const for tree-shaking
+// eslint-disable-next-line @typescript-eslint/no-redeclare
+export type DateFormat = (typeof DateFormat)[keyof typeof DateFormat]
