+++ packages/components/src/Calendar/index.ts
@@ -2,3 +2,4 @@ export * from "./CalendarSingle"
 export * from "./CalendarRange"
 export * from "./enums"
 export * from "./types"
+export * from "./utils"
