+++ packages/components/src/Calendar/utils/formatDateAsText.ts
@@ -1,5 +1,6 @@
 import { format } from "date-fns"
-import { DateFormat, DisabledDays } from "~components/Calendar"
+import { DateFormat } from "../enums"
+import { DisabledDays } from "../types"
 import { isDisabledDate } from "./isDisabledDate"
 import { isInvalidDate } from "./isInvalidDate"
