+++ packages/components/src/Calendar/utils/index.ts
@@ -1,9 +1,11 @@
 export * from "./calculateDisabledDays"
 export * from "./formatDateAsNumeral"
 export * from "./formatDateAsText"
-export * from "./getLocale"
 export * from "./isDisabledDate"
 export * from "./isInvalidDate"
+export * from "./isValidWeekStartsOn"
+export * from "./isSelectingDayInCalendar"
 export * from "./parseDateAsTextOrNumeral"
 export * from "./parseDateFromNumeralFormatValue"
 export * from "./parseDateFromTextFormatValue"
+export * from "./setFocusInCalendar"
