+++ packages/components/src/Calendar/utils/isSelectingDayInCalendar.spec.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 import { render, screen } from "@testing-library/react"
-import { CalendarSingle } from "~components/Calendar"
+import { CalendarSingle } from "../CalendarSingle"
 import { isSelectingDayInCalendar } from "./isSelectingDayInCalendar"
 
 describe("isSelectingDayInCalendar", () => {
