+++ packages/components/src/DatePicker/index.ts
@@ -1,3 +1,2 @@
 export * from "./DatePicker"
-export * from "../utils/date-controls/enums"
 export * from "./types"
