+++ packages/components/src/DatePicker/utils/getLocale.spec.ts
