+++ packages/components/src/DatePicker/utils/validateDate.spec.ts
@@ -1,4 +1,4 @@
-import { validateDate } from "../utils/validateDate"
+import { validateDate } from "./validateDate"
 
 describe("validateDate", () => {
   it("returns expected response when selected day is undefined", () => {
