+++ packages/components/src/DatePicker/utils/validateDate.ts
@@ -1,5 +1,8 @@
-import { DisabledDays } from "~components/Calendar"
-import { isDisabledDate, isInvalidDate } from "~utils/date-controls"
+import {
+  DisabledDays,
+  isDisabledDate,
+  isInvalidDate,
+} from "~components/Calendar"
 import { ValidationResponse } from "../types"
 
 export type ValidateDateArgs = {
